requirement.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.6.15.2
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
crispy==0.6.3
crispy-bootstrap5==0.6
cssselect==1.2.0
cycler==0.11.0
Django==4.1
django-crispy-forms==1.14.0
exceptiongroup==1.0.0
fabio==0.14.0
fake-useragent==0.1.11
fonttools==4.37.1
h11==0.14.0
h5py==3.7.0
idna==3.4
imap-tools==0.57.0
importlib-metadata==5.0.0
kiwisolver==1.4.4
kombu==5.2.4
lxml==4.9.1
matplotlib==3.5.3
nest-asyncio==1.5.6
numpy==1.23.2
outcome==1.2.0
packaging==21.3
pandas==1.4.3
parse==1.19.0
Pillow==9.2.0
plotly==5.10.0
prompt-toolkit==3.0.36
psycopg2==2.9.3
pycparser==2.21
pyee==8.2.2
pyparsing==3.0.9
pyppeteer==1.0.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.4.0
requests==2.28.1
selenium==4.6.0
silx==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tenacity==8.1.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.2
urllib3==1.26.12
v==0.0.0
vine==5.0.0
w3lib==2.0.1
wcwidth==0.2.5
webdriver-manager==3.8.4
websockets==10.4
wsproto==1.2.0
xvfbwrapper==0.2.9
zipp==3.10.0
